requirements.txt
litellm>=1.40.0
python-dotenv>=1.0.0
pyyaml>=6.0.1
aiohttp>=3.9.0
